pyproject.toml
[tool.poetry]
name = "hackernews-rag"
version = "0.1.0"
description = "RAG system for HackerNews content using Ollama"
authors = ["Oussama Anadani <anadani@student.chalmers.se>"]
readme = "README.md"
packages = [
    {include = "src"},
]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.35.0"
python-dotenv = "^1.1.0"
aiohttp = "^3.11.18"
langchain = "^0.3.25"
ijson = "^3.4.0"
langchain-community = "^0.3.24"
sentence-transformers = "^4.1.0"
protobuf = "3.20.3"
torch = "^2.7.0"
langchain-ollama = "^0.3.3"
chromadb = "^1.0.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pre-commit = "^4.2.0"
ruff = "^0.11.11"

[tool.poetry.scripts]
hackernews-pipeline = "src.run_pipeline:main"
hackernews-gui = "gui.streamlit_app:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
